task7 - controllers logic separated from routes
