+++ task7/src/routes/autoSuggest/index.js
@@ -1,32 +1,15 @@
-import express from 'express';
-
+import { Router } from 'express';
 import validators from './validators';
-import UserService from '../../services/userService';
-import db from '../../models';
-
 import { apiLoggerMiddleware as loggerMiddleware } from '../../utils/logger';
-import { populateErrorWithHadledFields } from '../../utils/errorHandler';
+import { getAutoSuggestedUsers } from '../../controllers/autoSuggestController';
 
-const userService = new UserService(db);
-const router = express.Router();
+const router = Router();
 
 router.get(
     '/getAutoSuggestUsers',
     validators.getAutoSuggestUsersQuery,
     loggerMiddleware,
-    async (req, res, next) => {
-        try {
-            const { limit, loginSubstring } = req.query;
-            const data = await userService.getFilteredSliceOfUsers(
-                limit,
-                loginSubstring
-            );
-            res.json(data);
-            return next();
-        } catch (err) {
-            return next(populateErrorWithHadledFields(err, req));
-        }
-    }
+    getAutoSuggestedUsers
 );
 
 export default router;
