+++ task7/src/routes/login/index.js
@@ -1,41 +1,10 @@
-import express from 'express';
-import jwt from 'jsonwebtoken';
-import UserService from '../../services/userService';
-import db from '../../models';
+import { Router } from 'express';
 import { apiLoggerMiddleware as loggerMiddleware } from '../../utils/logger';
-import {
-    ErrorHandler,
-    populateErrorWithHadledFields
-} from '../../utils/errorHandler';
 import validators from './validators';
+import { retrieveJwtToken } from '../../controllers/loginController';
 
-const router = express.Router();
-const userService = new UserService(db);
+const router = Router();
 
-router
-    .route('/')
-    .post(validators.login, loggerMiddleware, async (req, res, next) => {
-        try {
-            const { username, password } = req.body;
-            const { id: userid } = await userService.getUserByNameAndPassword(username, password);
-            if (userid) {
-                const accessToken = jwt.sign(
-                    { username, userid },
-                    process.env.TOKEN_SECRET,
-                    { expiresIn: '1800s' }
-                );
-                res.json({ accessToken });
-            } else {
-                throw new ErrorHandler(
-                    401,
-                    'Unauthenticated - bad credentials'
-                );
-            }
-            return next();
-        } catch (err) {
-            return next(populateErrorWithHadledFields(err, req));
-        }
-    });
-
+router.route('/').post(validators.login, loggerMiddleware, retrieveJwtToken);
 
 export default router;
